requirements.txt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
monero==0.7.2
requests==2.31.0
python-dotenv==1.0.0